Several changes has been introduced.
The clock (PAIP 18.7 The Tournament Version of Othello) is working almost as in the common lisp code.

The reversi, get-move, and print-board function have been updated accordingly.

Functions time-string has been coded, plus a few ancillary functions and constants.

A new Exception has been introduced. GameOverException, written in Clojure sing the
gen-class construct, is a real java class.
The GameOverException brings a "mutable" map that is used to propagate the final game value returned
by the reversi function. The map is implemented as an atom.

Files changed:
	new file:   GameOverException.clj
	modified:   auxfns.clj
	modified:   constants.clj
	modified:   reversi.clj