Several references to the PAIP book has been added.
A few function moved around, in the files and between them.

The PAIP code has been completly transcripted, a few sanity check
must be run.

The round-robin function applied to the basic strategies ends with
results that appear in line with the statistics presented into PAIP.

The code compiles without errors or warnings using sbcl.