Added random-reversi-series and switch-strategies functions.
switch-strategies uses the global variable *move-number*,
it has been implemented as an atom. The move number is
shared by the reversi function, that increment it and the
switch-strategies one.

This design, close to the PAIP Common Lisp version, is not usable
if any kind of concurrency is planned.
To fix it the global variable has to be substituted by a loop/recur
variable local to the reversi function, and passed around.

A better approach could be to either:

 - have a function that calculate it based on the board state
 - develop a complete game structure with board, next to play,
   moves, clock, move number.