round-robin function tests have been made.
The report printing has been modified to calculate the percentage of wins.
The comparison with the common lisp (and the PAIP outcomes) seams to says
that the round-robin works fine (and so the tournament version of reversi).

The comparison between alpha-beta-searcher and minimax-searcher:

reversi> (round-robin
	  (list (minimax-searcher 4 count-difference)
		(minimax-searcher 4 weighted-squares)
		(minimax-searcher 4 modified-weighted-squares)
		random-strategy)
	  5 10
	  '(count-difference weighted modified-weighted random))
count-difference      20.00:  -----    2.0   00.0   10.0
weighted              38.33:    8.0  -----    5.0   10.0
modified-weighted     41.67:   10.0    5.0  -----   10.0
random                00.00:   00.0   00.0   00.0  -----
nil
reversi> (round-robin
	  (list (alpha-beta-searcher 4 count-difference)
		(alpha-beta-searcher 4 weighted-squares)
		(alpha-beta-searcher 4 modified-weighted-squares)
		random-strategy)
	  5 10
	  '(count-difference weighted modified-weighted random))
count-difference      15.83:  -----    2.0   00.0    7.5
weighted              37.50:    8.0  -----    5.0    9.5
modified-weighted     40.00:   10.0    5.0  -----    9.0
random                 6.67:    2.5    0.5    1.0  -----
nil
reversi>

highlight a "statistical" difference between minimax-searcher
and alpha-beta-searcher.
It should'nt. It is almost sure that it is a bug into the alpha-beta
function.